Extract duplicated server connectivity check
Moved duplicate `performServerConnectivityCheck` and `fetchServerConnectivity` logic from `SignupActivity.kt` and `MyPlanetLite.kt` into a new `ServerConnectivityRepository`. Centralized `ServerConnectivityResult` into the shared `model` package.
