+++ app/src/main/java/org/ole/planet/myplanet/lite/model/ServerConnectivityResult.kt
@@ -1,0 +1,7 @@
+package org.ole.planet.myplanet.lite.model
+
+data class ServerConnectivityResult(
+    val reachable: Boolean,
+    val parentCode: String? = null,
+    val code: String? = null
+)
